fix bug when scale produces a float hitbox size for player and running animation is not working

+++ src/main/java/com/codecool/platformer/entities/Player.java
@@ -36,8 +36,8 @@ public class Player extends Entity {
         initHitbox(
                 x,
                 y,
-                SpriteSize.PLAYER.HITBOX_WIDTH * GameProperties.SCALE,
-                SpriteSize.PLAYER.HITBOX_HEIGHT * GameProperties.SCALE
+                (int) (SpriteSize.PLAYER.HITBOX_WIDTH * GameProperties.SCALE),
+                (int) (SpriteSize.PLAYER.HITBOX_HEIGHT * GameProperties.SCALE)
         );
     }
